Add Artist Search, Optimize DB Query

+++ backend/models/channel.py
@@ -5,7 +5,7 @@ class Channel(db.Model):
     __tablename__ = 'channels'
     
     id = db.Column(db.Integer, primary_key=True)
-    artist_id = db.Column(db.Integer, db.ForeignKey('Artists.id'), nullable=False)
+    artist_id = db.Column(db.BigInteger, db.ForeignKey('Artists.id'), nullable=False)
     platform = db.Column(db.Enum('instagram', 'youtube', 'tiktok', 'twitter'), nullable=False)
     channel_id = db.Column(db.String(100), nullable=False)
     channel_name = db.Column(db.String(200))
